Fix wrong size type in signal agent

+++ installer/pkgmgr_installer_signal_agent.c
@@ -223,7 +223,7 @@ static gboolean __handle_signal(gint fd, GIOCondition cond, gpointer user_data)
 	}
 
 	memcpy(&type_len, buf, sizeof(size_t));
-	memcpy(&data_len, buf + sizeof(int), sizeof(gsize));
+	memcpy(&data_len, buf + sizeof(size_t), sizeof(gsize));
 
 	r = recv(clifd, buf, type_len + data_len, 0);
 	if (r < 0) {
